Move files and modify to use parser
